fix HeaderClose cancellation
An appropriate header close amy be absent if the first character of
the rest of the string is anything but '=', not just a newline. For
instance `==abc==d` is an invalid header which would cause a runtime
panic with the old implementation, but not this one.

This patch also keeps the next state from relying on the whole string,
and just looks at the immediate next character.

In addition, this patch fixes an off-by-one.