build paragraph close one character at a time
This is roughly the same logic as fd1907c; we want to only analyze one
character at a time to simplify the state machine. To that end, we
introduce the state ParClose containing the string of the paragraph that
it's closing and an integer indicating how many newlines it has
encountered. If it's encountered two, it ends the paragraph, otherwise
it just collapses the whitespace and continues building the paragraph.

Behavior is unchanged.