rewrite performState to operate over a single character
Taking advantage of refactors 7163234, 110d6a2, and fd1907c, we rewrite
the performState function to have the following signature:

	performState :: State -> Char -> State

Aside from reducing the raw amount of and simplyifying the code, this
opens up the parser to future improvements in encapsulation and
idiomaticity, as it only relies on characters, not implementation
details of Strings, Texts, or ByteStrings. (Refactors exchanging one for
the other would probably want to move the gemtext function over to
main).

Note that this assumes that the acceptance of a token *always* should
always have the resumption of the parser start in LineStart, which,
while currently the case, won't hold with future additions.