wrangle newlines with ParagraphStart state
By being a bit more explicit about what's a paragraph and what's not, we
can ensure that newlines are managed a bit more prettily. If a newline
is to appear where a paragraph starts, it's just ignored. This forces
consistent behavior between, for instance, the following two samples of
wikitext:

	= Example 1 =
	A paragraph is started immediately.

	= Example 2 =

	The following paragraph is preceeded by a newline.

Both of these are now displayed identically as:

	# Example 1
	A paragraph is started immediately.

	# Example 2
	The following paragraph is preceeded by a newline.

In addition, this patch cleans up HorizontalRules such that they are
always surrounded by one newline before and after, mirroring what's
valid in wikitext[0]. (see the tiny change to the show implementation).

[0]: There's no standard, so this was found via trial and error in the
     Wikipedia's WYSIWYG editor: this may be incorrect, but at least it
     looks good :).